chore: rename barrel file
